GP2-1688: move contact-related helpers out of core and into contact app
