+++ core/helpers.py
@@ -404,23 +404,6 @@ def get_s3_file_stream(file_name, bucket_name=settings.AWS_STORAGE_BUCKET_NAME_D
     return s3_resource['Body'].read().decode('utf-8')
 
 
-def retrieve_regional_offices(postcode):
-    response = api_client.exporting.lookup_regional_offices_by_postcode(postcode)
-    response.raise_for_status()
-    return response.json()
-
-
-def retrieve_regional_office_email(postcode):
-    try:
-        office_details = retrieve_regional_offices(postcode)
-    except requests.exceptions.RequestException:
-        email = None
-    else:
-        matches = [office for office in office_details if office['is_match']]
-        email = matches[0]['email'] if matches else None
-    return email
-
-
 class GeoLocationRedirector:
     DOMESTIC_COUNTRY_CODES = ['GB', 'IE']
     COUNTRY_TO_LANGUAGE_MAP = {
